Initial move of canvas/context/snapshot size from u64 -> u32
Changed the two instances of `euclid::default::Size2D<u64>` in
`servo/components/script/canvas_context.rs` and
`servo/components/shared/snapshot/lib.rs` that were outlined as the bare minimum
in "Make canvas/context/snapshot size be u32 everywhere" to
`euclid::default::Size2D<u32>`. Every other change made in this commit is
either:
 - of similar nature, and is the minimum that would allow compilation
 - resolving lints triggered by the former

More might be needed to complete the issue, but I feel like this is a good
starting point.

This commit includes changes to the following components:
 - canvas
 - pixels
 - script
 - shared/canvas
 - shared/snapshot

+++ components/script/dom/htmlcanvaselement.rs
@@ -397,8 +397,8 @@ impl HTMLCanvasElement {
         // u32 can't panic, since the data comes from a canvas which is always smaller than
         // u32::MAX.
         let canvas_data = snapshot.data();
-        let width = snapshot.size().width as u32;
-        let height = snapshot.size().height as u32;
+        let width = snapshot.size().width;
+        let height = snapshot.size().height;
 
         match image_type {
             EncodedImageType::Png => {
